+++ JsonToMidiConverter/Models/Song/Beat.cs
@@ -22,15 +22,15 @@ public partial class Beat
     public bool Vibrato { get; set; }
     public Text? Text { get; set; }
     public bool LetRing { get; set; }
-    public int Dots { get; set; }
+    public byte Dots { get; set; }
     public bool Rest { get; set; }
     public bool Tapping { get; set; }
-    public int Tuplet { get; set; }
+    public byte Tuplet { get; set; }
     public bool TupletStart { get; set; }
     public bool TupletStop { get; set; }
     public GraceNote? GraceNote { get; set; }
-    public int UpStroke { get; set; }
-    public int DownStroke { get; set; }
+    public byte UpStroke { get; set; }
+    public byte DownStroke { get; set; }
 
     [JsonConverter(typeof(MarkerConverter))]
     public Marker? Chord { get; set; }
@@ -38,7 +38,7 @@ public partial class Beat
     public bool Popping { get; set; }
     public GradualVelocity? GradualVelocity { get; set; }
     public VibratoWithTremoloBar? VibratoWithTremoloBar { get; set; }
-    public int? VibratoBar { get; set; }
+    public byte? VibratoBar { get; set; }
 
 
     [JsonConverter(typeof(PickStrokeConverter))]
@@ -48,10 +48,10 @@ public partial class Beat
     public Bend? TremoloBar { get; set; }
     public bool WideVibrato { get; set; }
     public BrushStroke? BrushStroke { get; set; }
-    public int DownArpeggio { get; set; }
+    public byte DownArpeggio { get; set; }
     public bool HasRasgueado { get; set; }
     public BrushStroke? Arpeggio { get; set; }
-    public int UpArpeggio { get; set; }
+    public byte UpArpeggio { get; set; }
     public MeasureTempo? Tempo { get; set; }
     public bool Dotted { get; set; }
     public bool FadeIn { get; set; }
@@ -60,7 +60,7 @@ public partial class Beat
     public bool ArtificialHarmonic { get; set; }
     public bool PinchHarmonic { get; set; }
     public bool TapHarmonic { get; set; }
-    public int? WideVibratoBar { get; set; }
+    public byte? WideVibratoBar { get; set; }
 
     public bool DoubleDotted { get; set; }
