+++ JsonToMidiConverter/Models/Song/MusicalFraction.cs
@@ -1,6 +1,6 @@
 ﻿namespace JsonToMidiConverter.Models.Song;
 
-public record MusicalFraction(long Numerator, long Denominator)
+public record MusicalFraction(byte Numerator, byte Denominator)
 {
     public MusicalFraction Copy() => new(Numerator, Denominator);
 }
